New color map and markers for line plot
The color map chosen for the line plot made the lines difficult to see
in black and white because the lightest colour was too light. Fixed by
switching to a darker color map. I also made my own selection of markers
since the default selection had a few markers that were difficult to
tell apart for highly optimized plots.